requirements.txt
flask
flask-cors
python-dotenv
pandas
plotly
google-generativeai
python-dateutil
pymongo